updated simple if-else and loops to single-line format

+++ code/agent/core/ant.py
@@ -18,7 +18,7 @@
 #     E-Mail: rene.rath@growautomation.at
 #     Web: https://git.growautomation.at
 
-#ga_version0.3
+# ga_version 0.3
 
 from datetime import datetime
 from datetime import timedelta
@@ -29,9 +29,15 @@ from string import ascii_letters as string_ascii_letters
 from string import digits as string_digits
 from random import choice as random_choice
 from colorama import Fore as colorama_fore
+from inspect import getfile as inspect_getfile
+from inspect import currentframe as inspect_currentframe
 # from functools import lru_cache
 
 from ga.core.config import GetConfig
+from ga.core.smallant import SmallLogWrite
+
+
+SmallLogWrite("Current module: %s" % inspect_getfile(inspect_currentframe()), loglevel=2)
 
 
 # Just vars
@@ -40,19 +46,12 @@ log_redirect = " 2>&1 | tee -a %s" % GetConfig("path_log")
 # Time formats
 
 
-def now(format):
-    return datetime.now().strftime(format)
+def now(time_format):
+    return datetime.now().strftime(time_format)
 
 
-time01 = now("%H-%M-%S")
-time02 = now("%H:%M:%S")
-time03 = now("%H-%M")
-
-date01 = now("%Y-%m-%d")
-date02 = now("%Y")
-date03 = now("%m")
-date04 = now("%d")
-
+time01, time02, time03 = now("%H-%M-%S"), now("%H:%M:%S"), now("%H-%M")
+date01, date02, date03, date04 = now("%Y-%m-%d"), now("%Y"), now("%m"), now("%d")
 timestamp = "%Y-%m-%d %H:%M:%S"
 
 
@@ -66,26 +65,15 @@ class ShellOutput(object):
         self.start()
 
     def start(self):
-        if self.font == "text":
-            self.text()
-        elif self.font == "head":
-            self.header()
+        self.header() if self.font == "head" else self.text()
 
     def header(self):
         shellhight, shellwidth = os_popen('stty size', 'r').read().split()
         print("\n%s\n%s\n%s\n" % (self.symbol * (int(shellwidth) - 1), self.output, self.symbol * (int(shellwidth) - 1)))
 
     def colors(self):
-        if self.style == "warn":
-            return colorama_fore.YELLOW
-        elif self.style == "info":
-            return colorama_fore.CYAN
-        elif self.style == "err":
-            return colorama_fore.RED
-        elif self.style == "succ":
-            return colorama_fore.GREEN
-        else:
-            return ""
+        return colorama_fore.YELLOW if self.style == "warn" else colorama_fore.CYAN if self.style == "info" else colorama_fore.RED if self.style == "err" \
+            else colorama_fore.GREEN if self.style == "succ" else ""
 
     def text(self):
         print(self.colors() + "%s\n" % self.output + colorama_fore.RESET)
@@ -99,16 +87,11 @@ class LogWrite(object):
         self.log_level = level
 
     def __repr__(self):
-        if self.log_level > GetConfig("log_level"):
-            return False
-        else:
-            self.write()
+        return False if self.log_level > GetConfig("log_level") else self.write()
 
     def open(self):
         logdir = "%s/%s/%s" % (GetConfig("path_log"), self.scripttype, date02)
-        if os_path.exists(logdir) is False:
-            os_system("mkdir -p " + logdir)
-        return open("%s/%s_%s.log" % (logdir, date03, self.scripttype), 'a')
+        return os_system("mkdir -p " + logdir) if os_path.exists(logdir) is False else open("%s/%s_%s.log" % (logdir, date03, self.scripttype), 'a')
 
     def write(self):
         logfile = self.open()
@@ -129,79 +112,43 @@ class Line(object):
         self.backup = backup
 
     def __repr__(self):
-        if self.action == "find":
-            self.find()
-        elif self.action == "delete":
-            self.delete()
-        elif self.action == "replace":
-            self.replace()
-        elif self.action == "add":
-            self.add()
+        self.find() if self.action == "find" else self.delete() if self.action == "delete" else self.replace() if self.action == "replace" else self.add() if self.action == "add" else None
 
     def find(self):
         tmpfile = open(self.file, 'r')
-        for xline in tmpfile.readlines():
-            if xline.find(self.searchfor) != -1:
-                return xline
-        return False
+        for xline in tmpfile.readlines(): return xline if xline.find(self.searchfor) != -1 else False
 
     def delete(self):
-        if self.backup == "yes":
-            os_system("sed -i%s '/%s/d' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.file, self.file, self.backupfile, self.backupdir))
-        else:
-            os_system("sed -i '/%s/d' %s" % (self.searchfor, self.file))
+        os_system("sed -i%s '/%s/d' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.file, self.file, self.backupfile, self.backupdir)) if self.backup == "yes" \
+            else os_system("sed -i '/%s/d' %s" % (self.searchfor, self.file))
 
     def replace(self):
-        if self.backup == "yes":
-            os_system("sed -i%s 's/%s/%s/p' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.replacewith, self.file, self.file, self.backupfile, self.backupdir))
-        else:
-            os_system("sed -i 's/%s/%s/p' %s" % (self.searchfor, self.replacewith, self.file))
+        os_system("sed -i%s 's/%s/%s/p' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.replacewith, self.file, self.file, self.backupfile, self.backupdir)) if self.backup == "yes" \
+            else os_system("sed -i 's/%s/%s/p' %s" % (self.searchfor, self.replacewith, self.file))
 
     def add(self):
         # insert after linenr / search = linenr
-        if self.backup == "yes":
-            os_system("sed -i%s '%s a %s' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.replacewith, self.file, self.file, self.backupfile, self.backupdir))
-        else:
-            os_system("sed -i '%s a %s' %s" % (self.searchfor, self.replacewith, self.file))
+        os_system("sed -i%s '%s a %s' %s && mv %s %s %s" % (self.backupfile, self.searchfor, self.replacewith, self.file, self.file, self.backupfile, self.backupdir)) if self.backup == "yes" \
+            else os_system("sed -i '%s a %s' %s" % (self.searchfor, self.replacewith, self.file))
 
 
 def ga_setup_pwd_gen(stringlength):
-    chars = string_ascii_letters + string_digits + "!#-_"
-    return ''.join(random_choice(chars) for i in range(stringlength))
+    return ''.join(random_choice(string_ascii_letters + string_digits + "!#-_") for i in range(stringlength))
 
 
 # Searches nested keys for values -> gives back the name of the nested keys
 def dict_nested_search(dictionary, tosearch):
-    for key in dictionary:
-        for subkey in dictionary[key]:
-            if tosearch in subkey:
-                return subkey
-    return None
+    return [(subkey if tosearch in subkey else None) for key in dictionary for subkey in dictionary[key]]
 
 
 def string_check(string, maxlength=10, minlength=2):
-    char_blacklist = "!%$§?^´`µ{}()°><|\\*ÄÖÜüöä@,"
-    if type(string) != "str":
-        return False
-    elif len(string) > maxlength or len(string) < minlength:
-        return False
-    elif any((char in char_blacklist) for char in string):
-        return False
-    else:
-        return True
+    return False if type(string) != "str" else False if len(string) > maxlength or len(string) < minlength else False if any((char in "!%$§?^´`µ{}()°><|\\*ÄÖÜüöä@,") for char in string) else True
 
 
 def dict_keycheck(dictionary, dictkey):
-    if dictionary[dictkey] is None:
-        return False
-    elif dictkey in dictionary:
-        return True
-    else:
-        return False
+    return False if dictionary[dictkey] is None else True if dictkey in dictionary else False
 
 
 def time_subtract(subtract, timeformat=timestamp, both=False):
     calculated = (datetime.now() - timedelta(seconds=subtract)).strftime(timeformat)
-    if both is True:
-        return datetime.now().strftime(timeformat), calculated
-    return calculated
+    return datetime.now().strftime(timeformat), calculated if both is True else calculated
